+++ R/global-actions.R
@@ -91,12 +91,10 @@ globalevent_tool <- function(
 #' @rdname epiworldR-deprecated
 globalaction_tool <- function(...) {
 
-  .Deprecated(
+  .Defunct(
     new = "globalevent_tool"
-    )  
+    )
 
-  globalevent_tool(...)
-  
 }
 
 #' @export
@@ -134,7 +132,7 @@ globalevent_tool_logit <- function(
 #' @rdname epiworldR-deprecated
 globalaction_tool_logit <- function(...) {
 
-  .Deprecated(
+  .Defunct(
     new = "globalevent_tool_logit"
     )  
 
@@ -172,7 +170,8 @@ globalevent_set_params <- function(
 #' @rdname epiworldR-deprecated
 globalaction_set_params <- function(...) {
 
-  .Deprecated(
+
+  .Defunct(
     new = "globalevent_set_params"
     )  
 
@@ -234,7 +233,7 @@ globalevent_fun <- function(
 #' @rdname epiworldR-deprecated
 globalaction_fun <- function(...) {
 
-  .Deprecated(
+  .Defunct(
     new = "globalevent_fun"
     )